requirements.txt
requests>=2.26.0
psycopg2-binary>=2.9.1
tqdm>=4.62.3
colorama>=0.4.4
python-dotenv>=0.19.1
pytest>=6.2.5